move Rating and State enums into their own modules
