+++ fsrs/rating.py
@@ -1,0 +1,15 @@
+from enum import IntEnum
+
+
+class Rating(IntEnum):
+    """
+    Enum representing the four possible ratings when reviewing a card.
+    """
+
+    Again = 1
+    Hard = 2
+    Good = 3
+    Easy = 4
+
+
+__all__ = ["Rating"]
