+++ fsrs/state.py
@@ -1,0 +1,14 @@
+from enum import IntEnum
+
+
+class State(IntEnum):
+    """
+    Enum representing the learning state of a Card object.
+    """
+
+    Learning = 1
+    Review = 2
+    Relearning = 3
+
+
+__all__ = ["State"]
